+++ lib/features/search/sources/search_sources.dart
@@ -8,6 +8,7 @@ import 'fantlab_source.dart';
 import 'google_books_source.dart';
 import 'hardcover_source.dart';
 import 'igdb_games_source.dart';
+import 'kitsu_anime_source.dart';
 import 'kitsu_manga_source.dart';
 import 'mangabaka_source.dart';
 import 'mangadex_source.dart';
@@ -40,6 +41,7 @@ final List<SearchSource> searchSources = List<SearchSource>.unmodifiable(
     // MangaDex
     MangaDexSource(),
     // Kitsu
+    KitsuAnimeSource(),
     KitsuMangaSource(),
     // VNDB
     VndbSource(),